some refactorring
